fix 11467 - re add Amallyn Shadowguide
Amallyn Shadowguide was removed in
c694ca1dd3e07b62ed5673192c1f209b548bfd60 as part of the update to the 9E
Aeldari Codex. This unit should still be legal to take for Biel Tan
Craftworld rosters.

Changes untested locally. This is just a best-guess reversal of the
changes that removed her.